requirements.txt
catboost
cmdstanpy
holidays
kafka-python
numpy
pandas
plotly
polars
PyYAML
scikit-learn
setuptools-git 
streamlit